+++ docs/changelog.rst
@@ -17,8 +17,10 @@ Version 0.5
 - :class:`~authlib.client.OAuth2Session` supports more grant types.
 - Deprecate built-in cache. Read more on `issue#23`_.
 - **Redesigned OAuth 1 Flask server**. Read the docs :ref:`flask_oauth1_server`.
+- Deprecate ``client_model``. Read more on `issue#27`_.
 
 .. _`issue#23`: https://github.com/lepture/authlib/issues/23
+.. _`issue#27`: https://github.com/lepture/authlib/issues/27
 
 .. admonition:: Rollback